requirement.txt
django
mysqlclient
django-apscheduler
calc_work_hours